# release-22.2: tochar: implement cache for parse format