--- a/pkg/util/tochar/BUILD.bazel
+++ b/pkg/util/tochar/BUILD.bazel
@@ -4,6 +4,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "tochar",
     srcs = [
+        "cache.go",
         "constants.go",
         "tochar.go",
     ],
@@ -12,7 +13,9 @@ go_library(
     deps = [
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
+        "//pkg/util/cache",
         "//pkg/util/duration",
+        "//pkg/util/syncutil",
         "//pkg/util/timeutil/pgdate",
         "@com_github_cockroachdb_errors//:errors",
     ],
@@ -20,7 +23,10 @@ go_library(
 
 go_test(
     name = "tochar_test",
-    srcs = ["tochar_test.go"],
+    srcs = [
+        "cache_test.go",
+        "tochar_test.go",
+    ],
     args = ["-test.timeout=295s"],
     data = glob(["testdata/**"]),
     embed = [":tochar"],