--- a/pkg/util/tochar/tochar.go
+++ b/pkg/util/tochar/tochar.go
@@ -25,14 +25,13 @@ import (
 )
 
 // TimeToChar converts a time and a `to_char` format string to a string.
-func TimeToChar(t time.Time, f string) (string, error) {
-	return timeToChar(timeWrapper{t}, parseFormat(f))
+func TimeToChar(t time.Time, c *FormatCache, f string) (string, error) {
+	return timeToChar(timeWrapper{t}, c.lookup(f))
 }
 
 // DurationToChar converts a duration and a `to_char` format string to a string.
-func DurationToChar(d duration.Duration, f string) (string, error) {
-	// TODO(#sql-experience): consider caching parse formats.
-	return timeToChar(makeToCharDuration(d), parseFormat(f))
+func DurationToChar(d duration.Duration, c *FormatCache, f string) (string, error) {
+	return timeToChar(makeToCharDuration(d), c.lookup(f))
 }
 
 // timeInterface is intended as a pass through to timeToChar.