Backport 1/1 commits from #91564.

/cc @cockroachdb/release

Release justification: major perf improvement for new feature

---

Implemented a basic cache for parse formats, similar to `ReCache`. Haven't done any memory monitoring as the space this takes is predictable - cache size * maxCacheKeySize * some constant for vars.

Before:
```
BenchmarkTimeToChar-10        	  761224	      1577 ns/op
BenchmarkIntervalToChar-10    	  552865	      2226 ns/op
```

After:
```
BenchmarkTimeToChar-10        	 1719128	       702.8 ns/op
BenchmarkIntervalToChar-10    	 1551093	       771.4 ns/op
```

Epic: None

Release note (sql change, performance improvement): `to_char` now has caching for parse formats, marking a speedup when running to_char with the same format between sessions.